Debug functions moved to separate file
